Enhance RCV logic
Makes it spend fewer cycles waiting around and walking needlessly

It will now spend any energy is managed to grab from a source before emptying it before moving to the next source.

Additionally, it will drop energy is has gathered for other creeps if it cannot find anything to do with it. This sets up a small hand-off like step where the creep acts like a harvester gathering energy and dropping it to give the room a boost before the RCV dies.