Cap maxCreepsInRole in new room
This will prevent the upgraders from flooding the room and competing for the energy that should probably be used to build extensions instead.

New rooms aren't capable of making the L2 harvester yet, and so arr very scarce on energy.

This change helps to ensure all the bases are covered without over-saturating on any single role.